Swapped in checkip file
